requirements.txt
mysql-connector-python
python-dotenv
flask
werkzeug
pandas
openpyxl
